backend/requirements.txt
aerich==0.7.1
asyncmy==0.2.5
passlib==1.7.4
fastapi==0.88.0
python-jose==3.3.0
python-multipart==0.0.5
tortoise-orm==0.19.2
uvicorn==0.20.0